Fixing unit test after the test schema changes
git-svn-id: http://svn.code.sf.net/p/schemacrawler/code@285 3243f2e7-4e1b-42d2-9742-d984016aa316

+++ schemacrawler/src/test/java/schemacrawler/test/SchemaCrawlerTest.java
@@ -113,11 +113,11 @@ public class SchemaCrawlerTest
         }, {
             "INTEGER", "VARCHAR", "VARCHAR"
         }, {
-            "INTEGER", "INTEGER", "DECIMAL"
+            "INTEGER", "INTEGER", "FLOAT"
         }, {
-            "INTEGER", "INTEGER", "INTEGER", "INTEGER", "DECIMAL"
+            "INTEGER", "INTEGER", "INTEGER", "INTEGER", "FLOAT"
         }, {
-            "INTEGER", "VARCHAR", "DECIMAL"
+            "INTEGER", "VARCHAR", "FLOAT"
         }, {
             "INTEGER", "VARCHAR"
         }
